+++ src/tcerror.h
@@ -115,7 +115,7 @@ DEFINE_TCError(SameModOfRecv, "method definition is only allowed at the same mod
                               "that `%s' type is defined");
 DEFINE_TCError(SameNameField, "cannot define method: `%s', since `%s' type has same name field");
 DEFINE_TCError(UndefinedSymbol, "cannot access undefined symbol: `%s'%s");
-DEFINE_TCError(UndefinedType, "undefined type: `%s'");
+DEFINE_TCError(UndefinedType, "undefined type: `%s'%s");
 DEFINE_TCError(UndefinedField, "cannot access undefined field: `%s' for `%s' type%s");
 DEFINE_TCError(UndefinedMethod, "cannot call undefined method: `%s' for `%s' type%s");
 DEFINE_TCError(UndefinedInit, "constructor is not defined in `%s' type");
@@ -164,6 +164,12 @@ TypeCheckError createTCErrorImpl(const Node &node, const char *kind, const char 
 template <typename T, typename... Arg, typename = base_of_t<T, TCError>>
 inline TypeCheckError createTCError(const Node &node, Arg &&...arg) {
   return createTCErrorImpl(node, T::kind, T::value, std::forward<Arg>(arg)...);
+}
+
+inline void addSuggestionSuffix(std::string &value, StringRef suggestion) {
+  value += ", did you mean `";
+  value += suggestion;
+  value += "' ?";
 }
 
 } // namespace ydsh